movie_genie/backend/requirements.txt
# Flask Backend Dependencies for Movie Genie

# Core Flask
flask>=3.0.0
flask-cors>=4.0.0

# Database (you already have SQLAlchemy setup)
flask-sqlalchemy>=3.1.0
sqlalchemy>=2.0.0

# Data processing (you already use these)
pandas>=2.0.0
numpy>=1.24.0

# ML/AI dependencies (you already have these)
torch>=2.0.0
scikit-learn>=1.3.0

# API utilities
marshmallow>=3.20.0
webargs>=8.3.0

# Development
python-dotenv>=1.0.0

# Optional: Caching (for production optimization)
flask-caching>=2.1.0

# Optional: API documentation
flasgger>=0.9.7